renames webpack-metadata to webpack-meta-config #8090
