+++ arches/install/arches-templates/project_name/webpack/webpack-meta-config.js
