+++ releases/7.0.0_draft.md
@@ -86,7 +86,7 @@ JavaScript:
         ```
 
 2. Copy the `arches/install/arches-templates/project_name/webpack/` directory from arches core directory to the root directory of the project ( it should be on the same level as `package.json`)
-    - in the project's `webpack/webpack-metadata.js` ensure that `PROJECT_ROOT_DIRECTORY` and `ARCHES_CORE_DIRECTORY` are pointing to the correct directories in the system.
+    - in the project's `webpack/webpack-meta-config.js` ensure that `PROJECT_ROOT_DIRECTORY` and `ARCHES_CORE_DIRECTORY` are pointing to the correct directories in the system.
 
 3. Copy `.eslintrc.js`, `.eslintignore`, `.babelrc`, `.browserslistrc` and `.stylelintrc.json` from arches-core to your project directory.
